Bypass authgear token-only restrictions
Reloading the application logs out the user when using jwt token only
mechanism in authgear so we punch a hole right through it by unsafely
storing the token in local storage. This is mostly due to free tier
limitations (authgear requires paying account to use custom domain stuff
which is required to use httpOnly cookies with samesite, which is the
ideal solution, to be done later, maybe.